Aggregate.Aggregate

Description

A maybe reusable definition of what is an Aggregate...

Documentation

class Aggregate a where

Definition of the Aggregate behaviors

Methods

aggregateId :: a -> UUID

the id of and Aggregate

version :: a -> Version

an aggregate has a version based on its last update.

source

Arguments

:: Maybe a
if you provide an Aggregate, this funtion will try to apply the event on it

-> [Event]
a list of Event

-> Either String a
an error message of the Aggregate updated.

a way to source the aggregate from a list of events

type Version = Int

A Version as a simple integer