Aggregate.PM.Project

Description

One of the most important Aggregate of this application. It manages the main project description, with its backlog, its team and its planned workload.

Documentation

data Project

A project is the aggragation of...

type ProjectId = UUID

data ProjectMode

Constructors

Scrum Int

Kanban

genProjectId :: IO UUID

createProject

Arguments

:: ProjectId
the id of the project, provided by the command handler

-> String
the code name of the project (imutable)

-> Text
the fullName of the project

-> Day
the first day of the project

-> ProjectMode
choose the mode of the project.

-> UTCTime
the time of the event... Not very interresting

-> (Event, Project)

Used to create a Project

updateName

Arguments

:: Project
the project to update

-> Text
the new FullName of the project

-> UTCTime
the time of the update

-> (Event, Project)
the Domain Event and the updated project

Update the Name