Aggregate.PM.Types

Description

Contains the definitions of the common types used in our current project.

Documentation

type Price = Int

Price allowes to describe the cost of a ressource.

data AggregateType

The types of Aggregates used in our project

Constructors

ExternalUser
Users are external aggregates provided by an other micro-service.

Project
Project are managed by our micro-service

Imputation
Used to describe the work of an ExternalUser on the Project or other cases...