Command.Command

Description

A maybe reusable definition of what is a Command Handler...

Documentation

data CommandError

A command error... we have to work the definition of the error code.

class Command cmd where

A Command is ...

Methods

doFilter :: cmd -> Maybe CommandError

can be used to perform simple check on the command values (eg. empty string, ...)

execute

Arguments

:: Repository repo

=> cmd
the command

-> repo
the repository where the aggregate can be found

-> UTCTime
the time of the execution of the command

-> IO (Either CommandError ([Event], Version, UUID))
a command error or a t-uplet with a list of the domain event and the new version of the aggregate.

execute the command